+++ src/app/home/components/PostContainer/FilterContainer/FilterContainer.tsx
@@ -1,8 +1,8 @@
+import AreaFilter from './AreaFilter/AreaFilter';
+import { verticalLine } from './AreaFilter/AreaFilter.css';
+import ContactTargetFilter from './ContactTargetPopover/ContactTargetFilter';
 import { filterContainerLayout, resetFilterButton } from './FilterContainer.css';
-import AreaFilter from '../AreaFilter/AreaFilter';
-import { verticalLine } from '../AreaFilter/AreaFilter.css';
-import ContactTargetFilter from '../ContactTargetPopover/ContactTargetFilter';
-import MatchTypeFilter from '../MatchTypeFilter/MatchTypeFilter';
+import MatchTypeFilter from './MatchTypeFilter/MatchTypeFilter';
 
 import { ExperimentPostListFilters } from '@/apis/post';
 import Icon from '@/components/Icon';
